# Pr5SP/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Country": {
      "get": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_GetCountries",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_PostCountry",
        "requestBody": {
          "x-name": "country",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          }
        }
      }
    },
    "/api/Country/{id}": {
      "get": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_GetCountry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Country"
        ],
        "summary": "ааа",
        "operationId": "Country_PutCountry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "country",
          "description": "ааа",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Country"
        ],
        "operationId": "Country_DeleteCountry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Exhibit": {
      "get": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_GetExhibits",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exhibit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_PostExhibit",
        "requestBody": {
          "x-name": "exhibit",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exhibit"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibit"
                }
              }
            }
          }
        }
      }
    },
    "/api/Exhibit/{id}": {
      "get": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_GetExhibit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exhibit"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_PutExhibit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "exhibit",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exhibit"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_DeleteExhibit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pavilion": {
      "get": {
        "tags": [
          "Pavilion"
        ],
        "operationId": "Pavilion_GetPavilions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pavilion"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pavilion"
        ],
        "operationId": "Pavilion_PostPavilion",
        "requestBody": {
          "x-name": "pavilion",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pavilion"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pavilion"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pavilion/{id}": {
      "get": {
        "tags": [
          "Pavilion"
        ],
        "operationId": "Pavilion_GetPavilion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pavilion"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pavilion"
        ],
        "operationId": "Pavilion_PutPavilion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "pavilion",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pavilion"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pavilion"
        ],
        "operationId": "Pavilion_DeletePavilion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Country": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "countryId": {
            "type": "integer",
            "format": "int32"
          },
          "nameOfCountry": {
            "type": "string"
          },
          "exhibits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exhibit"
            }
          }
        }
      },
      "Exhibit": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exhibitId": {
            "type": "integer",
            "format": "int32"
          },
          "nameOfExhibit": {
            "type": "string"
          },
          "typeOfExhibit": {
            "type": "integer",
            "format": "int32"
          },
          "countryId": {
            "type": "integer",
            "format": "int32"
          },
          "dateOfArrival": {
            "type": "string",
            "format": "date"
          },
          "pavilionId": {
            "type": "integer",
            "format": "int32"
          },
          "discoveryDate": {
            "type": "string",
            "format": "date"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "pavilion": {
            "$ref": "#/components/schemas/Pavilion"
          },
          "typeOfExhibitNavigation": {
            "$ref": "#/components/schemas/TypeOfExhibit"
          }
        }
      },
      "Pavilion": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pavilionId": {
            "type": "integer",
            "format": "int32"
          },
          "nameOfPavilion": {
            "type": "string"
          },
          "responsibleEmployee": {
            "type": "integer",
            "format": "int32"
          },
          "exhibits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exhibit"
            }
          },
          "responsibleEmployeeNavigation": {
            "$ref": "#/components/schemas/ResponsibleEmployee"
          }
        }
      },
      "ResponsibleEmployee": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "responsibleEmployeeId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "patronymic": {
            "type": "string"
          },
          "telephoneNumber": {
            "type": "string"
          },
          "pavilions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pavilion"
            }
          }
        }
      },
      "TypeOfExhibit": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "typeOfExhibitId": {
            "type": "integer",
            "format": "int32"
          },
          "nameOfType": {
            "type": "string"
          },
          "exhibits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exhibit"
            }
          }
        }
      }
    }
  }
}